Add files via upload
### Repository Updates 🗓️

This repository will be regularly updated with new exercises each week as I progress through my Machine Learning Applications course. At the end of the term, a final project will be added, showcasing what I've learned and applying it to a comprehensive problem. Stay tuned for new content and follow along on this learning journey!